Changed return value of deleteBy from Long to int

+++ src/main/kotlin/com/grupo14IngSis/snippetSearcherAccessManager/repository/OwnershipsRepository.kt
@@ -20,19 +20,19 @@ interface OwnershipsRepository : JpaRepository<Ownerships, String> {
     fun deleteByOwnerIdAndSnippetId(
         @Param("ownerId") ownerId: String,
         @Param("snippetId") snippetId: String,
-    ): Long
+    ): Int
 
     @Transactional
     @Modifying
     @Query("DELETE FROM Ownerships o WHERE o.ownerId = :ownerId")
     fun deleteByOwnerId(
         @Param("ownerId") ownerId: String,
-    ): Long
+    ): Int
 
     @Transactional
     @Modifying
     @Query("DELETE FROM Ownerships o WHERE o.snippetId = :snippetId")
     fun deleteBySnippetId(
         @Param("snippetId") snippetId: String,
-    ): Long
+    ): Int
 }
